# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Search RAG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  filename: string;
  chunkCount: number;
  createdAt: string;
}

export interface SearchRequest {
  /** @minLength 1 */
  query: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  topK?: number;
}

export interface TextFragment {
  id: number;
  bookId: number;
  bookTitle: string;
  bookAuthor: string;
  chunkIndex: number;
  content: string;
  chapterTitle?: string;
  similarity: number;
}

export interface SearchFragmentsResponse {
  query: string;
  fragments: TextFragment[];
  found: boolean;
  message?: string;
}

export interface AnswerResponse {
  query: string;
  answer: string;
  citations: TextFragment[];
  found: boolean;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type UploadBookBody = {
  file: Blob;
  title?: string;
  author?: string;
};