# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Search RAG API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all uploaded books
 */
export const ListBooksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  filename: zod.string(),
  chunkCount: zod.number(),
  createdAt: zod.date(),
});
export const ListBooksResponse = zod.array(ListBooksResponseItem);

/**
 * @summary Upload a new book
 */
export const UploadBookBody = zod.object({
  file: zod.instanceof(File),
  title: zod.string().optional(),
  author: zod.string().optional(),
});

/**
 * @summary Delete a book
 */
export const DeleteBookParams = zod.object({
  bookId: zod.coerce.number(),
});

export const DeleteBookResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Search for relevant text fragments
 */

export const searchFragmentsBodyTopKDefault = 5;
export const searchFragmentsBodyTopKMax = 10;

export const SearchFragmentsBody = zod.object({
  query: zod.string().min(1),
  topK: zod
    .number()
    .min(1)
    .max(searchFragmentsBodyTopKMax)
    .default(searchFragmentsBodyTopKDefault),
});

export const SearchFragmentsResponse = zod.object({
  query: zod.string(),
  fragments: zod.array(
    zod.object({
      id: zod.number(),
      bookId: zod.number(),
      bookTitle: zod.string(),
      bookAuthor: zod.string(),
      chunkIndex: zod.number(),
      content: zod.string(),
      chapterTitle: zod.string().optional(),
      similarity: zod.number(),
    }),
  ),
  found: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Answer a question based on book content
 */

export const answerQuestionBodyTopKDefault = 5;
export const answerQuestionBodyTopKMax = 10;

export const AnswerQuestionBody = zod.object({
  query: zod.string().min(1),
  topK: zod
    .number()
    .min(1)
    .max(answerQuestionBodyTopKMax)
    .default(answerQuestionBodyTopKDefault),
});

export const AnswerQuestionResponse = zod.object({
  query: zod.string(),
  answer: zod.string(),
  citations: zod.array(
    zod.object({
      id: zod.number(),
      bookId: zod.number(),
      bookTitle: zod.string(),
      bookAuthor: zod.string(),
      chunkIndex: zod.number(),
      content: zod.string(),
      chapterTitle: zod.string().optional(),
      similarity: zod.number(),
    }),
  ),
  found: zod.boolean(),
  message: zod.string().optional(),
});