# lib/api-zod/src/generated/types/answerResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Search RAG API
 * OpenAPI spec version: 0.1.0
 */
import type { TextFragment } from "./textFragment";

export interface AnswerResponse {
  query: string;
  answer: string;
  citations: TextFragment[];
  found: boolean;
  message?: string;
}