# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Search RAG API
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  id: number;
  title: string;
  author: string;
  filename: string;
  chunkCount: number;
  createdAt: Date;
}