# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Search RAG API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchRequest {
  /** @minLength 1 */
  query: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  topK?: number;
}