# lib/api-zod/src/generated/types/textFragment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Search RAG API
 * OpenAPI spec version: 0.1.0
 */

export interface TextFragment {
  id: number;
  bookId: number;
  bookTitle: string;
  bookAuthor: string;
  chunkIndex: number;
  content: string;
  chapterTitle?: string;
  similarity: number;
}